Main Startup Function Done
I have the main startup class completed. It is called `StartPrinting` and is in the main package, `com.is2300.rcp`. The `public static void main(String[] args)` function parses the command line switches and arguments (targets) via an instance of `CmdLineParser` class, from the `ISCommandLineParser` library. We are allowing the following switches/targets, thus far:

- `-h | --help`: Shows some simple terminal help text and exits.
- `-v | --version`: Shows the program name, copyright information and version, then exits.
- `-f path | --folder path`: Provides the program the top-level folder to search through for files to print.

Once we get the printing working, we may want to accept file extension(s) to print. I'm considering doing this now in order to avoid refactoring later.

***NOTE:***
Command line switch/target parsing is working as expected. All tests I have thrown at it have succeeded. The only thing I haven't done is to try to get/create a POJO for our printing class, as I'm not sure we need to do that. As of right now, I've got the package `com.is2300.rcp.printer` package, with the `Printer` class set up with a `private` constructor and a single method: `public static boolean print(File toPrint)`. This method will process all of the printing and recursion.