Recursion Working
I have managed to get the recursion working as expected, but haven't yet limited it to only specified files. However, I have created thirty-nine (39) `FileFilter` extension classes for different programming languages, from Bash to XML and a whole lot in between. The various versions of Java/JS and C/C++/CSharp/ObjectiveC, to BASIC and Visual Basic, and including languages such as Ada and Lua and Perl. I think we've pretty much covered the gambit, but we can always add more.

Anyways, I'm committing and pushing to take a break...