Recursion Logic Error Fixed : Moving on to Actually Printing
Currently, I've just got the recursion working, after fixing a logic error that was preventing the printing loop when using a `FileFilter` for a specific language. I had the following for the logic in the `FileFilter` classes:

```java
    @Override
    public boolean accept(File file) {
        for ( String extension : extensions ) {
            if ( file.getName().toLowerCase().endsWith(extension) ) {
                return true;
            }
        }
        
        return false;
    }
```

But, when I'd run the program providing the `FileFilter`, it would not print any filenames to the screen (which is all that's happening right now). I kept stepping through the program and when I'd get to the `for` loop in this method:

```java
    public static boolean print(String pathToFiles, FileFilter filter) {
        if ( pathToFiles == null ) {
            throw new NullPointerException("The path cannot be null.");
        }
        
        boolean printed = false;
        File file = new File(pathToFiles);
        
        if ( file.isDirectory() ) {
            for ( File f : file.listFiles(filter) ) {
                if ( f.isDirectory() ) {
                    print(f.getAbsolutePath(), filter);
                } else {
                    printed = printFile(f);
                    printCount++;
                }
            }
        } else { 
            printed = printFile(file);
        }
        
        return printed;
    }
```

With my breakpoint set on the line `for ( File f : file.listFiles(filter) ) {`, I would check the `file.isDirectory()` return value and it would be true. Then when I'd check the `file.getPath()` return value, it would be the path that I set on the command line. Everything seemed correct, however, when I'd check the `file.filePath` variable value, it would be `null`. I couldn't figure out what was going on, so I took a little break and thought about it, then consulted the Java API JavaDoc for `FileFilter`. That's when it hit me that I had a logic flaw in the overridden `boolean accept(File file)` method. So, I added an `else` block to that method and it became:

```java
    @Override
    public boolean accept(File file) {
        for ( String extension : extensions ) {
            if ( file.getName().toLowerCase().endsWith(extension) ) {
                return true;
            } else {
                return file.isDirectory();
            }
        }
        
        return false;
    }
```

That `else` block returning `true` if the file is a directory did the trick. Now, the program recurses properly **and** prints only the files I want it to print.

Another issue that I ran into was with the Integrity Solutions CmdLineParser library...Even though I added it to my libraries folder in the project tree, when I built the Jar file and tried to execute it on the command line, I kept getting a `ClassNotFoundException`. Therefore, since the library is only a single class, I went ahead and created a new package in the project tree for `com.is2300.cmdlineparser` and just added that class into it, so the "whole" library (of one class :grin:) is included in the project source. Since doing this, I am able to execute the Jar on the command line the same way I do in the IDE.