Do Over
I'm trying to redo the `CmdLineParser` and place all arguments into a `java.util.Map` class, using the switches as the key and a list of options, up to the next switch, as the value. I get it close to working, but then it bombs out.

I've included comments of pseudocode of my thought processes. Need to take a look at this later.