Command Line Parsing Complete
Committing due to command line parsing now working! The `CmdLineParser` class now takes the arguments from the command line and parses them as follows:

* If the element of the argument array starts with "-":
    - the element is assigned to the variable `key` for later use
* Each element until the next element that starts with a "-":
    - added to the `List` variable `argList`
* The next time an element begins with a "-":
    - the `key` and `argList` are put into the `switches` `Map`
    - `key` is reassigned with the current element that starts with a "-"
    - `argList` is cleared of all entries
* Process repeats until the last element is reached.
* Once the last element is reached, the `else` clause is skipped, so I do a check on a `counter` variable that has been incremented with each iteration through the arguments array:
    - if the `counter` is equal to the length of the arguments array:
        * the `key` and `argList` are put to the `switches` `Map`.

Apparently, I was way over-thinking it at first, as I was getting some complex algorithms going and was not succeeding with what I was trying to do. A couple of Hershey's Kisses (aka "brain food") later, and a little break from the computer, and ***VOILA***, we've got a working parser again.

Now, all arguments to the program are stored in a `HashMap` with the switch as the key and all following arguments as the options to that switch. In order to use the values of the switch, one simply needs to:

```java
String[] values = cmdLineParserRef.getSwitches().getValue(key);
```

Due to storing the switches, options and targets in this manner, each switch may have multiple options, if needed. If the switch only has one expected option, then you may do the following to get its value:

```java
String value = cmdLineParserRef.getSwitches().getValue(key)[0];
```

The above *should* return only the 0th element of the array stored in the value of the key. This brings us one step closer to solving the feature request in the Issues section.