Printing Begun
I've gotten the project refactored and have begun work on printing. However, the first test could have been better...

So that I didn't use a lot of paper in the testing process, I tested the recursive printing on the CLArgParser project tree. I set the command line arguments to `-l java -d /project/path/CLArgParser/src`. It recursed the tree just fine, but only printed one file. AND, that file was not printed well.

The reason behind it only printing one file was that I had the definition of my printer object at the top of the `print(File fileToPrint)` method, instead of at the point of use inside the loop, so the thrown exception was `javax.print.PrintException: already printing` which is because the print job was already running and had not been recreated by a new creation of the `Printer` class. Therefore, the JVM and OS thought the print job was already running.

This iteration of the project, I was using the `javax.print.SimpleDoc` class in an effort to avoid having to do a bunch of `Graphics2D` work, but I guess that work will have to be done. I'm going to dig into the `javax.print` package a little more to learn what I can about other styles of `Doc` classes that may be defined before I start working with the `Graphics2D` class...

I will post scans of the first print test in the comment to the pull request on the repository.