Printing Continues
In this iteration, I've got the pages breaking cleaner and not dropping lines between pages. However, there is still the issue with changing/setting the font/size and some words being cut off on the right margin. 

Currently, the top, left and right margins are at 1", approximately. The bottom margin is between 1/2 and 1/4 inch in height. We are going to want to allow the user to set the margins on the pages and to propagate those settings to the printout. I will add an issue for this and place the `core` label on it, as well as the `enhancement` label.

Once I get these basic printouts working the way I want, I will start working on the feature request for storing projects and only printing modified files.