Printing Continues
This commit was working to resolving Issue #14, but I have, thus far, failed. This issue needs to remain open for now.

I've gotten it to the best I can figure out. I've tried using `DocPrintJob`, `DocFlavor`, `Doc`, `PrintRequestAttributeSet` and `PrintServiceLookup`, but keep getting unhelpful exception messages from the API and cannot figure out how to get the printouts any better.

Recursion still works and the files get printed like the last iteration. But, I can't seem to get it any better than that.

I have it using `PrintRequestAttributeSet` and I am setting the attributes for `Copies(1)` and `Sides.DUPLEX`, but it still only prints on one side of the page. I'm at a loss for right now.

Since I'm stuck with what I've got for the printouts, I'm going to move onto working on the functionality for storing the Project Name and last print date, as well as working on the printing method to only print those files that have changed since the last print date. In other words, I'm moving onto implementing the Feature Request in Issue #6.