Working on Issue #6
Reading files and data from the disk and populating the "Select Project" list. Also, determined that we need to have three lines in the file:

* Last Print Date
* Path to top-level source folder
* Programming language code.

I've got the reading finished. I also have it working when the user selects a project from the list, it autopopulates the following:

* Project Name field
* Project Source Code Path field
* Selects the proper language in the Language combobox
* Updates the last print date with the value from the file

I have done no writing as yet and also have not implemented saving a new project out to file. Will work on it more later.