Trying to Wrap Lines
I'm attempting to provide for line wrapping, but I'm not being very successful. The line wrapping that works creates a worse printout than what I get with the generic `Printer` class. Even lines that have a hard line return in the input file are not broken at the end, and I'm doing nothing in code to get rid of the `\n` line return characters as I read in the file contents. I am, however, splitting the words based upon the space character, but the `\n` line terminators are seeming to disappear as well. 

I will figure this out, I'm just going to take a break from it for a short while.

On a positive note, the files no longer print if they haven't been modified since the last time the project was printed. So, that's a plus...